inttest: make dut more robust
I ran into some issues where the DUT wasn't coming up in a reasonable
amount of time and essentially a race condition between the integration
tests themselves and the DUT being ready for them.

This adds a quick (ha!) shell script for making sure we are seeing
hello's from the DUT before running the integration tests.