Create CallWrap() for multi root XML
In case the NETCONF server returns a multi-root XML in its body we are
currently unable to access them, only the first one is unmarshaled and
returned to the caller.

Wrap all the body's roots into a XML tag that will become the only one
body's root. This will allow the caller to recover all the body's root
of the NETCONF reply.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>